Update the JAX example to work with Flower 1.0  (#1331)
