+++ examples/quickstart_jax/server.py
@@ -1,9 +1,7 @@
-"""Flower server example."""
-
 import flwr as fl
 
 if __name__ == "__main__":
     fl.server.start_server(
         server_address="0.0.0.0:8080",
-        config={"num_rounds": 3},
+        config=fl.server.ServerConfig(num_rounds=3),
     )
